fixture-ize
