+++ test/esm.spec.js
@@ -26,18 +26,7 @@ test('esm default export', t => {
   execSync('npm install ' + polyfillDir);
   
   // Create a test file that uses the default export
-  fs.writeFileSync('esm-default.mjs', `
-    import dc from 'dc-polyfill';
-    
-    // Verify the default export has all the expected methods
-    console.assert(typeof dc.channel === 'function', 'dc.channel is a function');
-    console.assert(typeof dc.hasSubscribers === 'function', 'dc.hasSubscribers is a function');
-    console.assert(typeof dc.Channel === 'function', 'dc.Channel is a function/class');
-    
-    // Create a channel and ensure the API works
-    const ch = dc.channel('test-channel');
-    console.assert(ch instanceof dc.Channel, 'channel instance should be instance of Channel');
-  `);
+  fs.copyFileSync(`${polyfillDir}/test/fixtures/esm-default.mjs`, 'esm-default.mjs');
   
   t.doesNotThrow(() => execSync('node esm-default.mjs'));
   process.chdir(polyfillDir);
@@ -54,27 +43,9 @@ test('esm named exports', t => {
   execSync('npm install ' + polyfillDir);
   
   // Create a test file that imports named exports individually
+  fs.copyFileSync(`${polyfillDir}/test/fixtures/esm-named.mjs`, 'esm-named.mjs');
   fs.writeFileSync('esm-named.mjs', `
-    // Import named exports individually to test each one
-    import { channel } from 'dc-polyfill';
-    import { subscribe } from 'dc-polyfill';
-    import { unsubscribe } from 'dc-polyfill';
-    import { Channel } from 'dc-polyfill';
-    import { hasSubscribers } from 'dc-polyfill';
-    import { tracingChannel } from 'dc-polyfill';
-    
-    // Simple existence test for each function
-    console.assert(typeof channel === 'function', 'channel is a function');
-    console.assert(typeof subscribe === 'function', 'subscribe is a function');
-    console.assert(typeof unsubscribe === 'function', 'unsubscribe is a function');
-    console.assert(typeof hasSubscribers === 'function', 'hasSubscribers is a function');
-    console.assert(typeof tracingChannel === 'function', 'tracingChannel is a function');
-    console.assert(typeof Channel === 'function', 'Channel is a function');
-    
-    // Basic API test - not trying to be comprehensive
-    const ch = channel('test-channel');
-    const handler = () => {};
-    subscribe('test-channel', handler);
+
   `);
   
   t.doesNotThrow(() => execSync('node esm-named.mjs'));
